Move loss and timesteps in operations/
